polymarket-bot/requirements.txt
numpy
pandas
websockets
httpx
python-dotenv
py-clob-client
scikit-learn
